refactor scripts and imports
